requirements.txt
django-taggit
django-embed-video
